+++ nodeWiFi.h
@@ -1,6 +1,8 @@
 #ifndef NODEWIFI_H_
 #define NODEWIFI_H_
+
 #include <ESP8266WiFi.h>
+
 //Define PORT AND IP PARAMETERS
 
 #define PORT 6000
